WIP: separate default video viewer code from features needed for drawing tools
